Support having both ES 1.x indexes and ES 0.9x indexes running concurrently, which is very helpful for migrating an index.
Also, try being a bit smarter about when to sending a new mapping.

When applying narrowing date filters, filter all of the results, not just the facets.

Addd the ability to pass a dictionary to ES.

Pass back the search size even if it's zero.

Add the ability to not store everything.